feat(collectors): Add temperature collector with filtering
This commit introduces a new temperature collector that can gather temperature metrics from the system.

The collector can be configured with include and exclude glob patterns to filter the collected metrics. This is useful to avoid collecting a large number of useless or invalid metrics.

The temperature collector is disabled by default. A warning has been added to the configuration file to inform users about the potential issues and how to configure the collector.